requirements.txt
Django==4.2.0
django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
langchain>=0.0.300
langchain-experimental>=0.0.30
langchain-community>=0.0.10
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.8